+++ exit_app.rb
@@ -1,0 +1,6 @@
+module ExitApp
+  def exit_app
+    puts 'Thank you for using this app!'
+    exit
+  end
+end
